A lot of fixes and additions
This commit performs a lot of changes and stability corrections on the frontend.
It adds the API module to communicate with the backend server RESTfull API. It
also fixes the application routing, the form creation and the integration with
Bootstrap 3. There are more fixes to add, but these are long due and need to be
committed.